+++ src/migtd/src/driver/serial.rs
@@ -35,7 +35,7 @@ impl Timer for SerailTimer {
         timer::reset_timer()
     }
 
-    fn set_timeout(&self, timeout: u64) -> Option<u64> {
+    fn set_timeout(&self, timeout: u32) -> Option<u32> {
         timer::schedule_timeout(timeout)?;
 
         // enable the interrupt to accept the timeout event
